Update extensions.js
Hexa didn’t specify the command prompts in the BE port, as he himself mentioned it was made for Flux… that’s fine, we have our own branch now. Much more complicated, yet simplistic for conversions in both Decimal and Number. Designed for Endgame. Of course, there are some things that Hexa didn’t take into account a few Endgame-exclusive features, but that’s what I’m here for. REVAMP OF THE BE PORT IS A GO.